requirements.txt
asgiref==3.3.1
attrs==20.3.0
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
iniconfig==1.1.1
mysqlclient==2.0.1
packaging==20.4
parameterized==0.7.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytest-django==4.1.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
